requirements.txt
rich
scapy
requests
cryptography
pywebcopy
python-crontab
dnspython 
python-whois